@@ -1,24 +1,20 @@
 {
   "name": "sqlite3",
   "description": "Asynchronous, non-blocking SQLite3 bindings",
-  "version": "5.1.6",
+  "version": "5.1.7",
   "homepage": "https://github.com/TryGhost/node-sqlite3",
   "author": {
     "name": "Mapbox",
     "url": "https://mapbox.com/"
   },
   "binary": {
-    "module_name": "node_sqlite3",
-    "module_path": "./lib/binding/napi-v{napi_build_version}-{platform}-{libc}-{arch}",
-    "host": "https://github.com/TryGhost/node-sqlite3/releases/download/",
-    "remote_path": "v{version}",
-    "package_name": "napi-v{napi_build_version}-{platform}-{libc}-{arch}.tar.gz",
     "napi_versions": [
       3,
       6
     ]
   },
   "contributors": [
+    "Daniel Lockyer <hi@daniellockyer.com>",
     "Konstantin Käfer <mail@kkaefer.com>",
     "Dane Springmeyer <dane@mapbox.com>",
     "Will White <will@mapbox.com>",
@@ -48,14 +44,15 @@
     "url": "https://github.com/TryGhost/node-sqlite3.git"
   },
   "dependencies": {
-    "@mapbox/node-pre-gyp": "^1.0.0",
-    "node-addon-api": "^4.2.0",
+    "bindings": "^1.5.0",
+    "node-addon-api": "^7.0.0",
+    "prebuild-install": "^7.1.1",
     "tar": "^6.1.11"
   },
   "devDependencies": {
-    "eslint": "6.8.0",
-    "mocha": "7.2.0",
-    "node-pre-gyp-github": "1.4.4"
+    "eslint": "8.56.0",
+    "mocha": "10.2.0",
+    "prebuild": "12.1.0"
   },
   "peerDependencies": {
     "node-gyp": "8.x"
@@ -69,12 +66,11 @@
     "node-gyp": "8.x"
   },
   "scripts": {
-    "build": "node-pre-gyp build",
-    "build:debug": "node-pre-gyp build --debug",
-    "install": "node-pre-gyp install --fallback-to-build",
-    "pretest": "node test/support/createdb.js",
-    "test": "mocha -R spec --timeout 480000",
-    "pack": "node-pre-gyp package"
+    "install": "prebuild-install -r napi || node-gyp rebuild",
+    "prebuild": "prebuild --runtime napi --all --verbose",
+    "rebuild": "node-gyp rebuild",
+    "upload": "prebuild --verbose --prerelease",
+    "test": "node test/support/createdb.js && mocha -R spec --timeout 480000"
   },
   "license": "BSD-3-Clause",
   "keywords": [
