@@ -1,36 +1,40 @@
-const express = require('express');
-const router = express.Router();
+import express from "express";
+import { createProiect, getProiecte } from "../dataAccess/ProiectDA.js";
 
-const db = require('../../database');
+const studentRouter = express.Router();
 
 // Route to create a new project
-router.post('/createProject', (req, res) => {
-  // Logic to create a new project
+studentRouter.route('/createProject').post(async (req, res) => {
+    return res.json(await createProiect(req.body));
 });
 
-// Route to join an existing project
-router.post('/joinProject', (req, res) => {
-  // Logic to join an existing project
-});
+studentRouter.route('/allProjects').get(async (req, res) => {
+    return res.json(await getProiecte());
+})
 
-// Route to edit project details
-router.put('/editProject/:projectId', (req, res) => {
-  // Logic to edit project details
-});
+// // Route to join an existing project
+// router.post('/joinProject', (req, res) => {
+//   // Logic to join an existing project
+// });
 
-// Route to add deliverables to a project
-router.post('/addDeliverable/:projectId', (req, res) => {
-  // Logic to add deliverables to a project
-});
+// // Route to edit project details
+// router.put('/editProject/:projectId', (req, res) => {
+//   // Logic to edit project details
+// });
 
-// Route to edit deliverable details
-router.put('/editDeliverable/:deliverableId', (req, res) => {
-  // Logic to edit deliverable details
-});
+// // Route to add deliverables to a project
+// router.post('/addDeliverable/:projectId', (req, res) => {
+//   // Logic to add deliverables to a project
+// });
 
-// Route to add personal review
-router.post('/addOwnReview/:projectId', (req, res) => {
-  // Logic to add personal review
-});
+// // Route to edit deliverable details
+// router.put('/editDeliverable/:deliverableId', (req, res) => {
+//   // Logic to edit deliverable details
+// });
 
-module.exports = router;
+// // Route to add personal review
+// router.post('/addOwnReview/:projectId', (req, res) => {
+//   // Logic to add personal review
+// });
+
+export default studentRouter
