@@ -1,16 +1,16 @@
-const express = require('express');
-const router = express.Router();
+// const express = require('express');
+// const router = express.Router();
 
-const db = require('../../database');
+// const db = require('../../database');
 
-// Route to get project information
-router.get('/getProjects', (req, res) => {
-  // Logic to get project information
-});
+// // Route to get project information
+// router.get('/getProjects', (req, res) => {
+//   // Logic to get project information
+// });
 
-// Route to get reviews for a project
-router.get('/getReviewsForProject/:projectId', (req, res) => {
-  // Logic to get reviews for a project
-});
+// // Route to get reviews for a project
+// router.get('/getReviewsForProject/:projectId', (req, res) => {
+//   // Logic to get reviews for a project
+// });
 
-module.exports = router;
+// module.exports = router;
